VideoDuration 추가
duration Int에서 Long으로 수정

#103

+++ app/src/test/java/com/dogeby/tagplayer/domain/tag/GetCommonTagsFromVideosUseCaseTest.kt
@@ -36,7 +36,7 @@ class GetCommonTagsFromVideosUseCaseTest {
         val expectedCommonTags = tags.take(2)
         val videosWithTags = videoIds.map {
             VideoWithTags(
-                Video(it, it.toString(), it.toString(), it.toString(), it.toInt(), it.toString(), it),
+                Video(it, it.toString(), it.toString(), it.toString(), it, it.toString(), it),
                 tags.take(it.toInt().coerceAtLeast(2))
             )
         }
